Amend PR comments
